% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{methods}
\name{AllClasses}
\alias{AllClasses}
\alias{PaperDownloader}
\title{pmcERR S4 classes implementation in R}
\description{
PaperDownloader class is used to download papers given some entities.
}
\section{Slots}{

\describe{
\item{\code{entities}}{character vector of entities.}

\item{\code{papers}}{list of papers, will be filled when downloaded.}

\item{\code{papersDir}}{character indicating the path to locally store papers (can be
character() in order to dont store them).}
}}
\examples{
## Create a PaperDownloader to query for two entities relations
entities <- c('terminal differentiation', 'basal-like');
pprDldr <- PaperDownloader(entities=entities, papersDir='/tmp/pmcERR/');

}
\seealso{
\code{\link{methods-PaperDownloader}}
}

