% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Paper.R
\docType{methods}
\name{methods-Paper}
\alias{methods-Paper}
\alias{getFullPaper}
\alias{getFullPaper,Paper-method}
\alias{getFullPaper,Paper-method}
\alias{getSentences}
\alias{getSentences,Paper-method}
\alias{getSentences,Paper-method}
\title{Paper methods}
\usage{
"methods-Paper"(paper)

getFullPaper(paper)

\S4method{getFullPaper}{Paper}(paper)

getSentences(paper)

\S4method{getSentences}{Paper}(paper)
}
\arguments{
\item{paper}{a Paper object.}
}
\value{
string, or character vector.
}
\description{
Methods to work with a Paper.
\code{getFullPaper} returns the full paper as a string.
\code{getSentences} returns the full paper in a vector separating each
sentence.
}
\examples{
## Create a PaperDownloader to query for two entities relations
entities <- c('terminal differentiation', 'basal-like');
pprDldr <- PaperDownloader(entities=entities, papersDir='/tmp/pmcERR/');

## Get the first paper
papers <- getPapers(pprDldr, n=1);
paper <- papers[[1]];

getSentences(paper);

}
