% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-PaperDownloader.R
\docType{methods}
\name{methods-PaperDownloader}
\alias{downloadPapers}
\alias{downloadPapers,PaperDownloader-method}
\alias{downloadPapers,PaperDownloader-method,numeric}
\alias{getPapers}
\alias{getPapers,PaperDownloader-method}
\alias{getPapers,PaperDownloader-method,numeric}
\alias{getPapersIds}
\alias{getPapersIds,PaperDownloader-method}
\alias{methods-PaperDownloader}
\title{PaperDownloader methods}
\usage{
"methods-PaperDownloader"(paperDownloader)

getPapersIds(paperDownloader)

\S4method{getPapersIds}{PaperDownloader}(paperDownloader)

downloadPapers(paperDownloader, ...)

\S4method{downloadPapers}{PaperDownloader}(paperDownloader, n = -1)

getPapers(paperDownloader, ...)

\S4method{getPapers}{PaperDownloader}(paperDownloader, n = -1)
}
\arguments{
\item{paperDownloader}{a PaperDownloader object.}

\item{...}{not in use.}

\item{n}{(optional) numeric parameter.}
}
\value{
PaperDownloader object with updated values. Or a list of Paper
objects.
}
\description{
Methods to work with a PaperDownloader.
\code{getPapersIds} gets the number of papers possible to download that
relate the entities. Updates the PaperDownloader.
\code{downloadPapers} downloads the papers (n is the max desired to download).
\code{getPapers} gets the list of downloaded Paper objects (n is the max
desired to download).
}
\examples{
## Create a PaperDownloader to query for two entities relations
entities <- c('terminal differentiation', 'basal-like');
pprDldr <- PaperDownloader(entities=entities, papersDir='/tmp/pmcERR/');

## Get the number of papers possible to download, updating the pprDldr
pprDldr <- getPapersIds(pprDldr);

## Download the first 3 papers
pprDldr <- downloadPapers(pprDldr, n=3);

## Get the first 3 papers
papers <- getPapers(pprDldr, n=3);
length(papers);

}
\seealso{
\code{\link{AllClasses}}
}

