% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Relations.R
\docType{methods}
\name{methods-Relations}
\alias{getRelations}
\alias{getRelations,list,character}
\alias{getRelations,list,character-method}
\alias{methods-Relations}
\title{Relations methods}
\usage{
"methods-Relations"(papers, entities)

getRelations(papers, entities)

\S4method{getRelations}{list,character}(papers, entities)
}
\arguments{
\item{papers}{a list of Paper objects to search for relations.}

\item{entities}{a character vector of entities to query for.}
}
\value{
Relations objects.
}
\description{
Methods to work with Relations.
\code{getRelations} gets a Relations. Contains the papers that relate the
given entities with the relating sentences.
}
\examples{
## Create a PaperDownloader to query for two entities relations
entities <- c('terminal differentiation', 'basal-like');
pprDldr <- PaperDownloader(entities=entities, papersDir='/tmp/pmcERR/');

## Get the number of papers possible to download, updating the pprDldr
pprDldr <- getPapersIds(pprDldr);

## Download the first 3 papers
pprDldr <- downloadPapers(pprDldr, n=3);

## Get the first 3 papers
papers <- getPapers(pprDldr, n=3);

relations <- getRelations(papers, entities);

}

