% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-pmcERR.R
\docType{methods}
\name{pmcERR-GUI}
\alias{pmcERR-GUI}
\alias{pmcERR}
\title{pmcERR GUI}
\usage{
pmcERR()
}
\description{
pmcERR package shiny-based graphical user interface (GUI)
\code{pmcERR} starts the GUI.
}
\examples{
## Start the GUI
\dontrun{
pmcERR();
}

}
