% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcERRPackage.R
\docType{package}
\name{pmcERRPackage}
\alias{pmcERRPackage}
\alias{pmcERRPackage-package}
\title{pmcERR package}
\description{
PubMed Central Entity Relation Recognition.
The pmcERR package allows downloading the scientific articles of PubMed
Central given a list of entities. Then given the downloaded articles looks
for relations between these entities within them. In this way, giving a
list of articles that validate the relationship between entities.
}
\author{
Juan C. Rodriguez \email{jcrodriguez@bdmg.com.ar}, Franco M. Luque
and Elmer A. Fernandez \email{efernandez@bdmg.com.ar}
}
\keyword{Central,}
\keyword{Data}
\keyword{Entity}
\keyword{Mining}
\keyword{NLP,}
\keyword{PubMed}
\keyword{Recognition,}
\keyword{Relation}

